DPP-268 PoC in memory fan-out state before making it functional
