+++ ledger/participant-integration-api/src/main/scala/platform/store/completions/CompletionsDao.scala
@@ -1,0 +1,30 @@
+package com.daml.platform.store.completions
+
+import akka.NotUsed
+import akka.stream.scaladsl.Source
+import com.daml.ledger.ApplicationId
+import com.daml.ledger.api.v1.command_completion_service.CompletionStreamResponse
+import com.daml.ledger.participant.state.v1.Offset
+import com.daml.lf.data.Ref
+import com.daml.logging.LoggingContext
+import com.daml.platform.store.dao.CommandCompletionsTable.CompletionStreamResponseWithParties
+
+import scala.concurrent.Future
+
+trait CompletionsDao {
+  def getCommandCompletions(
+      startExclusive: Offset,
+      endInclusive: Offset,
+      applicationId: ApplicationId,
+      parties: Set[Ref.Party],
+  )(implicit
+      loggingContext: LoggingContext
+  ): Source[(Offset, CompletionStreamResponse), NotUsed]
+
+  def getAllCommandCompletions(
+      startExclusive: Offset,
+      endInclusive: Offset,
+  )(implicit
+      loggingContext: LoggingContext
+  ): Future[List[(Offset, CompletionStreamResponseWithParties)]]
+}
